requirements.txt
albucore==0.1.5
albumentationsx==2.2.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
click==8.3.3
contourpy==1.3.3
cycler==0.12.1
filelock==3.29.0
fonttools==4.62.1
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.12.0
idna==3.13
imbalanced-learn==0.14.1
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.9
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numkong==7.6.0
numpy==2.4.4
opencv-python-headless==4.13.0.92
packaging==26.2
pandas==3.0.2
pillow==12.2.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
rich==15.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
segmentation_models_pytorch==0.5.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.5
stringzilla==4.6.0
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.26
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
typer==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
